+++ src/extension-firefox/skipto.js
@@ -917,7 +917,8 @@ $skipToId-overlay .overlay-info.hasInfoBottom {
     menuBackgroundColor: '#dddddd',
   };
 
-  const MORE_INFO_URL='https://skipto-landmarks-headings.github.io/page-script-5/shortcuts.html';
+  const MORE_PAGE_INFO_URL='https://skipto-landmarks-headings.github.io/page-script-5/page.html';
+  const MORE_SHORTCUT_INFO_URL='https://skipto-landmarks-headings.github.io/page-script-5/shortcuts.html';
 
   const styleTemplate$1 = document.createElement('template');
   styleTemplate$1.textContent = `
@@ -977,6 +978,29 @@ dialog#skip-to-info-dialog .content {
   margin-right: 2em;
   margin-top: 0;
   margin-bottom: 2em;
+}
+
+dialog#skip-to-info-dialog .content .desc {
+  max-width: 20em;
+}
+
+dialog#skip-to-info-dialog .content .happy {
+  margin-top: 0.5em;
+  text-align: center;
+  font-family: fantasy, cursive;
+  font-size: 1.25em;
+  font-weight: bold;
+  font-style: italic;
+  letter-spacing: 0.05em;
+}
+
+
+dialog#skip-to-info-dialog .content .version,
+dialog#skip-to-info-dialog .content .copyright {
+  margin-top: 0.5em;
+  text-align: center;
+  font-weight: bold;
+  font-size: 90%;
 }
 
 dialog#skip-to-info-dialog .content table {
@@ -1045,7 +1069,7 @@ button:hover {
 }
 `;
 
-  class ShortcutsInfoDialog extends HTMLElement {
+  class SkipToContentInfoDialog extends HTMLElement {
     constructor () {
 
       super();
@@ -1090,6 +1114,8 @@ button:hover {
       this.closeButton2.addEventListener('click', this.onCloseButtonClick.bind(this));
       this.closeButton2.addEventListener('keydown', this.onKeyDown.bind(this));
 
+      this.moreInfoURL = '';
+
     }
 
     onCloseButtonClick () {
@@ -1102,7 +1128,9 @@ button:hover {
     }
 
     onMoreInfoClick () {
-      window.open(MORE_INFO_URL, '_blank').focus();
+      if (this.moreInfoURL) {
+        window.open(this.moreInfoURL, '_blank').focus();
+      }
     }
 
     configureStyle(config={}) {
@@ -1157,11 +1185,13 @@ button:hover {
     }
 
 
-    updateContent (config) {
+    updateShortcutContent (config) {
 
         while (this.contentElem.lastElementChild) {
           this.contentElem.removeChild(this.contentElem.lastElementChild);
         }
+
+        this.moreInfoURL = MORE_SHORTCUT_INFO_URL;
 
         this.h2Elem.textContent = config.shortcutsInfoLabel;
         this.closeButton1.setAttribute('aria-label', config.closeLabel);
@@ -1254,6 +1284,41 @@ button:hover {
         addRow(tbodyElem2, config.shortcutHeadingH4, config.msgH4Headings);
         addRow(tbodyElem2, config.shortcutHeadingH5, config.msgH5Headings);
         addRow(tbodyElem2, config.shortcutHeadingH6, config.msgH6Headings);
+
+    }
+
+    updateAboutContent (config) {
+
+      while (this.contentElem.lastElementChild) {
+        this.contentElem.removeChild(this.contentElem.lastElementChild);
+      }
+
+      this.moreInfoURL = MORE_PAGE_INFO_URL;
+
+      this.h2Elem.textContent = config.aboutInfoLabel;
+      this.closeButton1.setAttribute('aria-label', config.closeLabel);
+      this.closeButton2.textContent = config.closeLabel;
+      this.moreInfoButton.textContent = config.moreInfoLabel;
+
+      let divElem = document.createElement('div');
+      divElem.className = 'desc';
+      divElem.textContent = config.aboutDesc;
+      this.contentElem.appendChild(divElem);
+
+      divElem = document.createElement('div');
+      divElem.className = 'happy';
+      divElem.textContent = config.aboutHappy;
+      this.contentElem.appendChild(divElem);
+
+      divElem = document.createElement('div');
+      divElem.className = 'version';
+      divElem.textContent = config.aboutVersion;
+      this.contentElem.appendChild(divElem);
+
+      divElem = document.createElement('div');
+      divElem.className = 'copyright';
+      divElem.textContent = config.aboutCopyright;
+      this.contentElem.appendChild(divElem);
 
     }
 
@@ -3484,8 +3549,12 @@ div#skip-to-message.fade {
         this.shortcutsGroupNode.setAttribute('aria-labelledby', 'id-skip-to-menu-shortcutse-group-label');
         this.menuNode.appendChild(this.shortcutsGroupNode);
 
-        window.customElements.define("skip-to-shortcuts-info-dialog", ShortcutsInfoDialog);
-        this.infoDialog = document.createElement('skip-to-shortcuts-info-dialog');
+        if (this.config.aboutSupported === 'true') {
+          this.renderAboutToMenu(this.menuNode, this.config);
+        }
+
+        window.customElements.define("skip-to-content-info-dialog", SkipToContentInfoDialog);
+        this.infoDialog = document.createElement('skip-to-content-info-dialog');
         this.infoDialog.configureStyle(this.config);
         document.body.appendChild(this.infoDialog);
 
@@ -3721,6 +3790,33 @@ div#skip-to-message.fade {
       }
 
       /*
+       * @method renderAboutToMenu
+       *
+       * @desc Render the about menuitem
+       *
+       * @param  {Object}  menuNode   -  DOM element node for the menu
+       */
+      renderAboutToMenu (menuNode, config) {
+
+        const separatorNode = document.createElement('div');
+        separatorNode.setAttribute('role', 'separator');
+
+        const menuitemNode = document.createElement('div');
+        menuitemNode.setAttribute('role', 'menuitem');
+        menuitemNode.setAttribute('data-about-info', '');
+        menuitemNode.className = 'skip-to-nav skip-to-nesting-level-0';
+        menuitemNode.tabIndex = -1;
+
+        const labelNode = document.createElement('span');
+        labelNode.classList.add('label');
+        labelNode.textContent = config.aboutInfoLabel;
+        menuitemNode.appendChild(labelNode);
+
+        menuNode.appendChild(separatorNode);
+        menuNode.appendChild(menuitemNode);
+      }
+
+      /*
        * @method renderMenuitemToGroup
        *
        * @desc Renders a menuitem using an information object about the menuitem
@@ -3743,7 +3839,7 @@ div#skip-to-message.fade {
           menuitemNode.setAttribute('aria-label', mi.ariaLabel);
         }
 
-        // add event handlers
+        // add to group
         groupNode.appendChild(menuitemNode);
 
         // add heading level and label
@@ -4470,10 +4566,17 @@ div#skip-to-message.fade {
         }
 
         if (tgt.hasAttribute('data-shortcuts-info')) {
-          this.infoDialog.updateContent(this.skipToContentElem.config);
+          this.infoDialog.updateShortcutContent(this.skipToContentElem.config);
           this.infoDialog.openDialog();
           this.closePopup();
         }
+
+        if (tgt.hasAttribute('data-about-info')) {
+          this.infoDialog.updateAboutContent(this.skipToContentElem.config);
+          this.infoDialog.openDialog();
+          this.closePopup();
+        }
+
       }
 
       handleMenuitemKeydown(event) {
@@ -4689,14 +4792,14 @@ div#skip-to-message.fade {
   debug$1.flag = false;
 
 
-  class SkipToContent571 extends HTMLElement {
+  class SkipToContent572 extends HTMLElement {
 
     constructor() {
       // Always call super first in constructor
       super();
       this.attachShadow({ mode: 'open' });
       this.skipToId = 'id-skip-to';
-      this.version = "5.7.1";
+      this.version = "5.7.2";
       this.buttonSkipTo = false;
       this.initialized = false;
 
@@ -4748,6 +4851,13 @@ div#skip-to-message.fade {
         shortcutsToggleEnableLabel:  'Enable shortcuts',
         shortcutsToggleDisableLabel: 'Disable shortcuts',
         shortcutsInfoLabel:          'Shortcut Information',
+
+        aboutSupported: 'true',
+        aboutInfoLabel: `About SkipTo.js`,
+        aboutHappy: `Happy Skipping!`,
+        aboutVersion: `Version ${this.version}`,
+        aboutCopyright: 'BSD License, Copyright 2021-2025',
+        aboutDesc: 'SkipTo.js is a free and open source utility to support authors in implementing the WCAG 4.2.1 Bypass Block requirement on their websites.',
 
         closeLabel: 'Close',
         moreInfoLabel: 'More Information',
@@ -4831,7 +4941,8 @@ div#skip-to-message.fade {
         "data-skipto",
         "setfocus",
         "type",
-        "shortcuts"
+        "shortcuts",
+        "about"
         ];
     }
 
@@ -4853,6 +4964,15 @@ div#skip-to-message.fade {
         }
         else {
           this.config.shortcuts = 'disabled';
+        }
+      }
+
+      if (name === 'about') {
+        if (newValue.trim().toLowerCase() === 'true') {
+          this.config.aboutSupported = 'true';
+        }
+        else {
+          this.config.aboutSupported = 'false';
         }
       }
 
@@ -5135,7 +5255,7 @@ div#skip-to-message.fade {
           if (!isExtensionLoaded) {
             if (!isBookmarkletLoaded) {
               removePageSkipTo();
-              window.customElements.define(SkipToBookmarkletElmName, SkipToContent571);
+              window.customElements.define(SkipToBookmarkletElmName, SkipToContent572);
               skipToContentElem = document.createElement(SkipToBookmarkletElmName);
               skipToContentElem.setAttribute('version', skipToContentElem.version);
               skipToContentElem.setAttribute('type', type);
@@ -5151,10 +5271,11 @@ div#skip-to-message.fade {
           if (!isExtensionLoaded) {
             removePageSkipTo();
             removeBookmarkletSkipTo();
-            window.customElements.define(SkipToExtensionElmName, SkipToContent571);
+            window.customElements.define(SkipToExtensionElmName, SkipToContent572);
             skipToContentElem = document.createElement(SkipToExtensionElmName);
             skipToContentElem.setAttribute('version', skipToContentElem.version);
             skipToContentElem.setAttribute('type', type);
+            skipToContentElem.setAttribute('about', 'false');
             // always attach SkipToContent element to body
             if (document.body) {
               document.body.insertBefore(skipToContentElem, document.body.firstElementChild);
@@ -5164,7 +5285,7 @@ div#skip-to-message.fade {
 
         default:
           if (!isPageLoaded && !isBookmarkletLoaded && !isExtensionLoaded) {
-            window.customElements.define(SkipToPageElmName, SkipToContent571);
+            window.customElements.define(SkipToPageElmName, SkipToContent572);
             skipToContentElem = document.createElement(SkipToPageElmName);
             skipToContentElem.setAttribute('version', skipToContentElem.version);
             skipToContentElem.setAttribute('type', type);